move data to src
